requirements.txt
httpx[http2]==0.27.0
tenacity==8.3.0
rich==13.7.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
